Coding style suggestions (#213)
* Coding style suggestions.

* added param in/out

* Style suggestions for permutation tests.

* Some more doxygen formatting.

---------
