+++ tests/unit/hykkt/CMakeLists.txt
@@ -8,14 +8,14 @@
 
 # Build HyKKT tests
 
-add_executable(runHykktPermTests.exe runHykktPerm.cpp)
-target_link_libraries(runHykktPermTests.exe PRIVATE resolve_hykkt)
+add_executable(runHykktPermutationTests.exe runHykktPermutationTests.cpp)
+target_link_libraries(runHykktPermutationTests.exe PRIVATE resolve_hykkt)
 
 # Install tests
-set(installable_tests runHykktPermTests.exe)
+set(installable_tests runHykktPermutationTests.exe)
 
 install(TARGETS ${installable_tests}
         RUNTIME DESTINATION bin/resolve/tests/unit)
 
-add_test(NAME hykkt_perm_test   COMMAND $<TARGET_FILE:runHykktPermTests.exe>)
+add_test(NAME hykkt_perm_test   COMMAND $<TARGET_FILE:runHykktPermutationTests.exe>)
